Update rawData references to preLabeledData in protocol.yaml
Changed all occurrences of 'rawData' to 'preLabeledData', as part of repository redefinition. This was done to provide more explicit context and clarity on the data type being manipulated by each endpoint. Changes were made on the get, delete, and post endpoints, and in the summaries of these endpoints. The objective is to better reflect the current state of the 'knowledge map' ('知识图谱') project, which is now dealing with pre-annotated samples rather than raw samples.