Refactor API endpoints and update schema in protocol.yaml
Refactored the API endpoints in the knowledge graph's protocol.yaml file and updated the schema. The '/taskCreate/train' and '/taskCreate/imageGather' endpoints were renamed to '/rule/train/save' and '/rule/dataset/save' respectively to improve the clarity of their roles. Also, changes in the request body schemas of these endpoints were made to provide more detailed training and data-gathering instructions. It aims to enhance the expressiveness and accurateness of our API, making it more intuitive for the users.