Add descriptions to fields in protocol.yaml
The 'protocol.yaml' file was missing descriptions for several fields which could make understanding the purpose and function of each field difficult. Added descriptions provide insight into the function of each field and improve the readability of the file.

The following descriptions are added:

- AI应用信息 for cameraAiInfo
- AI应用ID for cameraAIId
- 采集间隔 for interval
- 告警目标阈值 for objectThresh
- 配置模型信息 for modelInfoList
- 模型优先级 for index
- 模型ID for modelId
- 目标标签列表 for objLabels
- 操作符 E 检测到所有目标 IN 检测到目标之一 for operator

These descriptions will help developers and users of this file better understand how it functions.