Update schema in protocol.yaml to allow task updates
In order to facilitate updates to training tasks, the 'ruleBody' schema in the 'protocol.yaml' has been revised. The 'name' and 'description' fields are not obligatory anymore. A 'ruleUpdateBody' schema has been introduced especially for aggregate training task updates, requiring an 'id' and defining important properties to capture necessary details for task updates.