Change data type of 'id' in protocol.yaml from string to integer
The commit changes the data type of 'id' in several schemas from string to integer. This change has been made to ensure that all 'id' fields in the 'protocol.yaml' have consistent data types. This improvement will help avoid potential type-related issues in the future and will bring uniformity to the codebase.