Add API endpoint and modify modelInfo attributes
Introduced a new '/model/list' endpoint to retrieve the list of models for the applications. Updated the modelInfo attributes to include more detailed information and removed some unnecessary elements. The adjustments were made based on the changed requirements and to allow a more precise and effective collection of model data.