Refactor protocol.yaml for better data structure and security
Updated protocol.yaml file to enforce the use of 'id' in 'requestBody' instead of query parameters and changed 'id' type to 'string'. This serves a better structure, flexibility, and security as it reduces the risk of parameter pollution attacks. Also, large chunks of commented code were eliminated for cleaner and more maintainable code.