Add AI knowledge event addition to protocol.yaml
A new post request path '/rule/event/save' was added to protocol.yaml file. This was to support the addition of AI knowledge events in the rule aggregator service. A new schema 'ruleEventBody' has also been added to define the structure of the request body for this new endpoint. This update enhances the functionality of the rule aggregator service by allowing addition of AI specific events.