Update API endpoint in YAML file
The API endpoint for obtaining the model list in the 'protocol.yaml' file has been changed. The previous endpoint '/model/list' has been updated to '/data-engine/model/list'. This change aims to improve the organization and make the endpoint more descriptive, indicating that it is part of the 'data-engine' processes.