Refactor API endpoint and 'protocol.yaml' schema
The API endpoint for fetching the model list was moved from '/data-engine/model/list' to '/data-engine/api/v1/data-engine/model/list'. The schema was also updated for clarity and organization. The previous term 'chipset' was replaced with 'chipType' and 'labelClassSetObjInfo' was changed from an object to an array. Additionally, a new property 'evolutionSetObjInfo' was introduced. These changes provide a better understanding of the data and ensure adherence to updated standards.